+++ tests/helpers/createBaseAdapterTests.ts
@@ -1,4 +1,4 @@
-import {  W3CStats as W3C } from '@observertc/sample-schemas-js';
+import * as W3C from '../../src/schema/W3cStatsIdentifiers';
 import { Adapter } from "../../src/adapters/Adapter";
 import { StatsValue } from "../../src/utils/StatsVisitor";
 import * as Generator from "./StatsGenerator";